서버 지형 다중 Landscape 지원 및 물리/스폰 로직 리팩토링
다만 아직 클라이언트단에 있는 물리 월드와 중력때문에 떨어지는중

MapDataManager가 여러 Landscape(지형 타일)을 지원하도록 구조를 대폭 변경하고, 각 타일별로 물리 Shape을 미리 생성/공유하도록 개선했습니다. Room에서는 모든 지형 타일에 대해 물리 바디를 생성·관리하며, 플레이어 스폰 시 월드 중앙에서 레이캐스트로 지면을 정확히 탐지해 항상 안전하게 스폰되도록 보정 로직을 추가했습니다. 높이 계산, 맵 경계 반환 등도 다중 타일을 고려하도록 일원화했으며, float 타입 일관성 및 불필요 코드 정리 등도 포함됩니다.

+++ Client/Client/MainPlayerScript.cpp
@@ -501,7 +501,7 @@ void MainPlayerScript::die_ui_update(float deltaTime)
 	UIManager::instance()->set_visible(UILayer::FRONT, "Death_UI", true);
 
 	// 실제 동작 -> update에서
-	alpha_background += static_cast<double>(deltaTime * 0.25);
+	alpha_background += deltaTime * 0.25f;
 	alpha_text += deltaTime * 0.25f * 0.5f;
 
 	alpha_background = std::min(alpha_background, 1.f);
